+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Home directories can now be created, changed and moved
 * Users can now be created for projects when updating them
+* You can now set or modify the default shell for users
 
 ### Changed
 * Login page is now styled to match dark mode